Removed unnecessary racist joke from readme
See issue https://github.com/aFarkas/html5shiv/issues/91

Using the word 'homie' in this way directly implies that PoC are violent and creates an association between them and improvised weapons.

jonathantneal has claimed that '"homie" means "a friend from back home"' as a defence of this joke, despite the fact that if that were true, and the intended meaning of the joke, it wouldn't actually make sense, since "friends from back home" have no link to improvised weapons. Clearly this explanation makes no sense, and the intended implication of the joke is that Crockford shares the qualities of those who would usually be called "homies" (if you're wondering, that's people of colour) and those qualities include a propensity towards violence.

Essentially there is really no reason for this joke to be present. It's not funny, and it's even not even a joke when read in the way Jonathan is pretending it was meant to be.