Remove unnecessary anonymous scope
Creating an annonymous scope requires creation of a function which can
only be done with a rather long ‘function’ keyword.  This isn’t helping
with minification so it’s better avoidid in possible.

Wrapping try-catch blocks in a function helps with engine’s optimisation
but HTML5 Shiv is hardly a piece of code which needs to be super-fast.
Furthermore, browsers which don’t support HTML5 are likely to not
benefit from the optimisation either.

Get rid of the anonymous function.